Get tmux to open new windows and panes in cwd
cwd-current working directory

The latest version of tmux (1.9) dropped that feature. The older version of tmux
(1.4) didn't have that feature either. Naturally, I was a little confused when
the tmux out on Linux didn't work of the box and an upgrade didn't help either.
I'm now including the officially recommended change in the tmux conf.