Better error checking and directory creation
`mkdir -p` is a better of way of creating directories only when they don't
already exist.

Additionally, I had to check for whether a non-directory file by the same
name as the intended directory already existed. If so, the script had to abort
after emitting an appropriate message.

I also made progress towards a more modular organization of code in the
shell scripts.