Making the install script more modular
It was hard to figure out which part of the script was doing what. Splitting up
the install script into functions was the logical thing to do.

In the process, I found some mistakes in the structure of the script related to
the creation of the 'install_dir'. Next step would be to fix that.