Fixing presence of globstar for older bash
When I allowed the install script to be run by an older bash, I didn't account
for the globstar because the older versions of bash didn't throw an error. But
the globstar (**) is not recognized by older versions of bash (< 4.0.0). This
led to some subtle errors in the detection of files based on globbed paths.

The change to `find` makes the install script compatible with older versions of
bash.